requirements.txt
pillow
pytesseract
matplotlib
